fix cmd folder
