+++ cmd/baton-successfactors/main.go
@@ -1,3 +1,6 @@
+================
+File: cmd/baton-successfactors/main.go
+================
 package main
 
 import (
@@ -9,9 +12,10 @@ import (
 	"github.com/conductorone/baton-sdk/pkg/connectorbuilder"
 	"github.com/conductorone/baton-sdk/pkg/field"
 	"github.com/conductorone/baton-sdk/pkg/types"
+	"github.com/conductorone/baton-successfactors/pkg/connector"
+	"github.com/conductorone/baton-successfactors/pkg/connector/client"
 	"github.com/grpc-ecosystem/go-grpc-middleware/logging/zap/ctxzap"
 	"github.com/spf13/viper"
-	"github.com/conductorone/baton-successfactors/pkg/connector"
 	"go.uber.org/zap"
 )
 
@@ -25,7 +29,8 @@ func main() {
 		"baton-successfactors",
 		getConnector,
 		field.Configuration{
-			Fields: ConfigurationFields,
+			Fields:           ConfigurationFields,
+			FieldRelationships: FieldRelationships,
 		},
 	)
 	if err != nil {
@@ -44,19 +49,36 @@ func main() {
 
 func getConnector(ctx context.Context, v *viper.Viper) (types.ConnectorServer, error) {
 	l := ctxzap.Extract(ctx)
-	if err := ValidateConfig(v); err != nil {
+
+	if err := ValidateConfig(ctx, v); err != nil {
 		return nil, err
 	}
 
-	cb, err := connector.New(ctx)
+	baseURL := v.GetString(BaseURLField.GetName())
+	clientID := v.GetString(ClientIDField.GetName())
+	companyID := v.GetString(CompanyIDField.GetName())
+	certPrivateKey := v.GetString(CertPrivateKeyField.GetName())
+	samlAssertionURL := v.GetString(SamlAssertionURLField.GetName())
+	idpIssuerURL := v.GetString(IdpIssuerURLField.GetName())
+	idpSubjectNameID := v.GetString(IdpSubjectNameIDField.GetName())
+
+	sfClient, err := client.New(ctx, baseURL, clientID, companyID, certPrivateKey, samlAssertionURL, idpIssuerURL, idpSubjectNameID)
+	if err != nil {
+		l.Error("error creating client", zap.Error(err))
+		return nil, err
+	}
+
+	cb, err := connector.New(ctx, sfClient)
 	if err != nil {
 		l.Error("error creating connector", zap.Error(err))
 		return nil, err
 	}
-	connector, err := connectorbuilder.NewConnector(ctx, cb)
+
+	c, err := connectorbuilder.NewConnector(ctx, cb)
 	if err != nil {
 		l.Error("error creating connector", zap.Error(err))
 		return nil, err
 	}
-	return connector, nil
+
+	return c, nil
 }
